Initialize color picker on load not on show
Prior to this change the color picker would end up at a wrong position,
on the first view. Apparently due to the fact that the HTML-contents are
being initialized during the first 'showPickerContent'.

The code might break some functionality, as the scope of variables has
changed to be global to webshim, instead to just one picker. From what I
gathered, these variables were used in a global sense anyways.